+++ NEWS.md
@@ -8,6 +8,7 @@ Version 0.19.1
  - Do not invalidate renderer caches when NO cache_buster is given (#158)
  - Call afterLayergroupCreate only after MapConfig verification (#159)
  - Drop LRU cache for "seen" layergroups (#160)
+ - Clearer error on mapnik tokens use with attribute service (#154)
 
  Bug fixes: